+++ Solution/RiskOfChaos/EffectDefinitions/Meta/EndAllEffects.cs
@@ -1,9 +1,12 @@
-﻿using RiskOfChaos.EffectHandling;
+﻿using RiskOfChaos.ConfigHandling;
+using RiskOfChaos.EffectHandling;
 using RiskOfChaos.EffectHandling.Controllers;
 using RiskOfChaos.EffectHandling.EffectClassAttributes;
+using RiskOfChaos.EffectHandling.EffectClassAttributes.Data;
 using RiskOfChaos.EffectHandling.EffectClassAttributes.Methods;
 using RiskOfChaos.EffectHandling.EffectComponents;
 using RiskOfChaos.Utilities.Extensions;
+using RiskOfOptions.OptionConfigs;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Networking;
@@ -13,6 +16,27 @@ namespace RiskOfChaos.EffectDefinitions.Meta
     [ChaosEffect("end_all_effects", DefaultSelectionWeight = 0.7f)]
     public class EndAllEffects : MonoBehaviour
     {
+        [EffectConfig]
+        static readonly ConfigHolder<bool> _excludePermanentEffects =
+            ConfigFactory<bool>.CreateConfig("Exclude Permanent Effects", true)
+                               .Description("If permanent effects should be excluded.")
+                               .OptionConfig(new CheckBoxConfig())
+                               .Build();
+
+        static bool canEndEffect(ChaosEffectComponent effectComponent, in EffectCanActivateContext context)
+        {
+            if (effectComponent.TryGetComponent(out ChaosEffectDurationComponent durationComponent))
+            {
+                if (durationComponent.TimedType == TimedEffectType.AlwaysActive)
+                    return false;
+
+                if (durationComponent.TimedType == TimedEffectType.Permanent && _excludePermanentEffects.Value)
+                    return false;
+            }
+
+            return effectComponent.IsRelevantForContext(context);
+        }
+
         [EffectCanActivate]
         static bool CanActivate(in EffectCanActivateContext context)
         {
@@ -20,10 +44,9 @@ namespace RiskOfChaos.EffectDefinitions.Meta
             {
                 foreach (ChaosEffectComponent effectComponent in ChaosEffectTracker.Instance.AllActiveTimedEffects)
                 {
-                    if (effectComponent.ChaosEffectInfo is TimedEffectInfo timedEffectInfo)
+                    if (canEndEffect(effectComponent, context))
                     {
-                        if (ChaosEffectTracker.Instance.IsAnyInstanceOfTimedEffectRelevantForContext(timedEffectInfo, context))
-                            return true;
+                        return true;
                     }
                 }
             }
@@ -46,19 +69,15 @@ namespace RiskOfChaos.EffectDefinitions.Meta
             List<ChaosEffectComponent> effectsToEnd = [];
             if (ChaosEffectTracker.Instance)
             {
+                EffectCanActivateContext activateContext = EffectCanActivateContext.Now;
+
                 effectsToEnd.EnsureCapacity(ChaosEffectTracker.Instance.AllActiveTimedEffects.Count);
                 foreach (ChaosEffectComponent effectComponent in ChaosEffectTracker.Instance.AllActiveTimedEffects)
                 {
-                    if (effectComponent == _effectComponent)
-                        continue;
-                    
-                    if (effectComponent.TryGetComponent(out ChaosEffectDurationComponent durationComponent))
+                    if (effectComponent != _effectComponent && canEndEffect(effectComponent, activateContext))
                     {
-                        if (durationComponent.TimedType == TimedEffectType.AlwaysActive)
-                            continue;
+                        effectsToEnd.Add(effectComponent);
                     }
-
-                    effectsToEnd.Add(effectComponent);
                 }
             }